TypeScript migration of framework.js
