+++ web/lite.html
@@ -17,7 +17,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1"><link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/5.0.0/normalize.min.css">
     <link rel="stylesheet" href="./css/style.css">
     <link rel="stylesheet" href="./css/themes.css">
-    <script type="module" src="./js/framework.js" defer></script>
+    <script type="module" src="./dist/framework.js" defer></script>
     <script type="module" src="./js/script.js" defer></script>
 
     <script type="text/javascript" defer>
@@ -361,7 +361,7 @@
       <div class="components">
         <div class="chip">
           <div class="text-container">
-            <p id="c-name-tag" class="c-name-tag" spellcheck="false" contenteditable>Enter Name...</p>
+            <input id="c-name-tag" class="c-name-tag" spellcheck="false" placeholder="Enter Name...">
           </div>
         </div>
